requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.5.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==0.19.2
PyMySQL==1.0.2
fake-useragent==1.4.0
bcrypt==3.2.0
numpy==1.26.3
plotly==5.18.0
python-dateutil==2.8.2
selenium==4.18.1
pillow==10.2.0
urllib3==2.2.1 